refactor(product, category): remove the product query manager
