--- libs/product/src/drivers/injection-tokens/product-query-manager.token.ts
@@ -1,8 +1,0 @@
-import { InjectionToken } from '@angular/core';
-import { DaffProductQueryManagerInterface } from '../interfaces/product-query-manager.interface';
-
-/**
- * Injection token that serves as a placeholder for any service that implements the DaffProductQueryManagerInterface.
- */
-export const DaffProductQueryManager = 
-    new InjectionToken<DaffProductQueryManagerInterface>('DaffProductQueryManager');
