+++ libs/product/src/drivers/magento/product-driver.module.ts
@@ -6,8 +6,6 @@ import { DaffMagentoProductService } from './product.service';
 
 import { DaffProductTransformer } from '../injection-tokens/product-transformer.token';
 import { DaffMagentoProductTransformerService } from './transforms/product-transformer.service';
-import { DaffProductQueryManager } from '../injection-tokens/product-query-manager.token';
-import { DaffMagentoProductGraphQlQueryManagerService } from './queries/product-query-manager.service';
 
 @NgModule({
   imports: [
@@ -26,10 +24,6 @@ export class DaffProductMagentoDriverModule {
         {
           provide: DaffProductTransformer,
           useExisting: DaffMagentoProductTransformerService
-        },
-        {
-          provide: DaffProductQueryManager,
-          useExisting: DaffMagentoProductGraphQlQueryManagerService
         }
       ]
     };
