+++ libs/product/src/drivers/magento/product.service.ts
@@ -8,10 +8,11 @@ import { Apollo } from 'apollo-angular';
 import { DaffProductServiceInterface } from '../interfaces/product-service.interface';
 import { DaffProductTransformer } from '../injection-tokens/product-transformer.token';
 import { DaffProductTransformerInterface } from '../interfaces/product-transformer.interface';
-import { DaffProductQueryManager } from '../injection-tokens/product-query-manager.token';
 import { DaffProductUnion } from '../../models/product-union';
-import { DaffMagentoProductQueryManagerInterface } from './interfaces/magento-product-query-manager.interface';
 import { DaffSortField } from './models/sort-field';
+import { GetSortFieldsAndFiltersByCategory } from './queries/get-sort-fields-and-filters-by-category';
+import { GetAllProductsQuery } from './queries/get-all-products';
+import { GetProductQuery } from './queries/get-product';
 
 /**
  * A service for making magento apollo queries for products of type, DaffProductUnion.
@@ -21,8 +22,7 @@ import { DaffSortField } from './models/sort-field';
 })
 export class DaffMagentoProductService implements DaffProductServiceInterface<DaffProductUnion> {  
   constructor(
-    private apollo: Apollo, 
-    @Inject(DaffProductQueryManager) public queryManager: DaffMagentoProductQueryManagerInterface,
+    private apollo: Apollo,
     @Inject(DaffProductTransformer) public transformer: DaffProductTransformerInterface<DaffProductUnion>) {}
 
   /**
@@ -30,7 +30,12 @@ export class DaffMagentoProductService implements DaffProductServiceInterface<Da
    * @param productId a product Id
    */
   get(productId: string): Observable<DaffProductUnion> {
-    return this.apollo.query<any>(this.queryManager.getAProductQuery(productId)).pipe(
+    return this.apollo.query<any>({
+			query: GetProductQuery,
+			variables: {
+				sku: productId
+			}
+		}).pipe(
       map(result => this.transformer.transform(result.data))
     );
   }
@@ -39,13 +44,20 @@ export class DaffMagentoProductService implements DaffProductServiceInterface<Da
    * Get an Observable of an array of DaffProductUnions.
    */
   getAll(): Observable<DaffProductUnion[]> {
-    return this.apollo.query<any>(this.queryManager.getAllProductsQuery()).pipe(
+    return this.apollo.query<any>({
+			query: GetAllProductsQuery
+		}).pipe(
       map(result => this.transformer.transformMany(result.data.products.items))
     );
   }
 
   getSortFieldsAndFiltersByCategory(categoryId: string): Observable<DaffSortField[]> {
-    return this.apollo.query<any>(this.queryManager.getSortFieldsAndFiltersByCategory(categoryId)).pipe(
+    return this.apollo.query<any>({
+			query: GetSortFieldsAndFiltersByCategory,
+			variables: {
+				categoryId: categoryId
+			}
+		}).pipe(
       map(result => result.data.products.sort_fields.options)
     );
   }
