+++ libs/product/src/drivers/magento/public_api.ts
@@ -2,6 +2,7 @@ export { ProductImageNode } from './models/product-image-node';
 export { ProductNode } from './models/product-node';
 export { ProductPriceNode } from './models/product-price-node';
 export { DaffMagentoProductTransformerService } from './transforms/product-transformer.service';
-export { DaffMagentoProductQueryManagerInterface } from './interfaces/magento-product-query-manager.interface';
 export { DaffSortField } from './models/sort-field';
-export { DaffMagentoProductGraphQlQueryManagerService } from './queries/product-query-manager.service';
+export { GetProductQuery } from './queries/get-product';
+export { GetAllProductsQuery } from './queries/get-all-products';
+export { GetSortFieldsAndFiltersByCategory } from './queries/get-sort-fields-and-filters-by-category';
