+++ libs/product/src/drivers/magento/queries/get-all-products.ts
@@ -1,0 +1,32 @@
+import gql from 'graphql-tag';
+
+export const GetAllProductsQuery = gql`
+query GetAllProducts($pageSize: Int)
+{
+	products(search: "Shirt", pageSize: $pageSize)
+	{
+		total_count
+		items {
+			id
+			name
+			sku
+			url_key
+			image {
+				url
+				label
+			}
+			price {
+				regularPrice {
+					amount {
+						value
+						currency
+					}
+				}
+			}
+		}
+		page_info {
+			page_size
+			current_page
+		}
+	}
+}`
